Update config.yml
Make the cost of pearls increase more often. Previously set to 365, Wing believed it to be set to 90, I am suggesting it be changed to 21. A period of 90 days means that it takes 2 years for a pearl to get to a place where it takes the daily essence of two players (not taking into account buying essence or stockpiling)